# docs/openapi.yaml
...
openapi: 3.0.1
info:
  title: SEGS API
  description: 'Secure Email Gateway Suite HTTP API. Authenticate with a stateful JWT (``Authorization:
    Bearer``) or the ``seg_session`` cookie. SCIM 2.0 is at ``/scim/v2``. Interactive ``/docs``
    is not served; the contract is ``docs/openapi.yaml`` and is what API Gateway imports.'
  version: 1.0.0
  x-generated-by: backend.api.openapi
servers:
- url: /
  description: Same origin as the SOC console (CloudFront)
paths:
...
  /api/sso-config:
    get:
      summary: Get Sso Config
      operationId: get_sso_config_api_sso_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - bearerAuth: []
      - cookieAuth: []
    put:
      summary: Set Sso Config
      operationId: set_sso_config_api_sso_config_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsoConfigUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
...
  /api/sender-profiles:
    get:
      summary: List Sender Profiles
      operationId: list_sender_profiles_api_sender_profiles_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
          maxLength: 200
          default: ''
          title: Q
      - name: ready
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Ready
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/sender-profiles/by-address:
    get:
      summary: Get Sender Profile
      operationId: get_sender_profile_api_sender_profiles_by_address_get
      parameters:
      - name: sender
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 320
          title: Sender
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SsoConfigUpdate:
      properties:
        enabled:
          type: boolean
          title: Enabled
          default: false
        issuer:
          type: string
          title: Issuer
          default: https://oauth.id.jumpcloud.com
        authorization_endpoint:
          type: string
          title: Authorization Endpoint
          default: ''
        token_endpoint:
          type: string
          title: Token Endpoint
          default: ''
        userinfo_endpoint:
          type: string
          title: Userinfo Endpoint
          default: ''
        client_id:
          type: string
          title: Client Id
          default: ''
        client_secret:
          type: string
          title: Client Secret
          default: ''
        allowed_domains:
          type: string
          title: Allowed Domains
          default: pdax.ph
        default_role:
          type: string
          enum:
          - viewer
          - analyst
          - admin
          title: Default Role
          default: viewer
      type: object
      title: SsoConfigUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Stateful JWT (RFC 6750). Same token as the seg_session cookie; jti is stored
        server-side.
    cookieAuth:
      type: apiKey
      in: cookie
      name: seg_session
      description: Session cookie set at login. Same credential as Bearer.
    scimBearer:
      type: http
      scheme: bearer
      description: SCIM 2.0 bearer (SEG_SCIM_BEARER_TOKEN) or an admin JWT.